refactor: remove repetition
