--- src/modules/CreateEntityPageContent/components/VideoContentCard/VideoContentCard.styles.ts
@@ -1,10 +1,0 @@
-import styled from 'styled-components'
-
-export const RemoveButton = styled.button`
-  border: none;
-  color: #39c3e6;
-  background: transparent;
-  font-size: 1rem;
-  font-weight: bold;
-  line-height: 1.2;
-`
