+++ src/modules/CreateEntityPageContent/components/VideoContentCard/VideoContentCard.tsx
@@ -6,7 +6,7 @@ import * as formUtils from '../../../../common/components/JsonForm/JsonForm.util
 import { FormData } from '../../../../common/components/JsonForm/types'
 import VideoLoader from '../../../../common/components/DropZone/FileLoader/FileLoader'
 import { FileType } from 'src/common/components/DropZone/types'
-import { RemoveButton } from './VideoContentCard.styles'
+import { RemoveButton } from '../PageContent.styles'
 
 interface Props {
   id: string
